Allow to specify another lockfile
