+++ pdm/cli/commands/lock.py
@@ -4,14 +4,14 @@ from pdm import signals
 from pdm.cli import actions
 from pdm.cli.commands.base import BaseCommand
 from pdm.cli.commands.run import run_script_if_present
-from pdm.cli.options import no_isolation_option
+from pdm.cli.options import lockfile_option, no_isolation_option
 from pdm.project import Project
 
 
 class Command(BaseCommand):
     """Resolve and lock dependencies"""
 
-    arguments = BaseCommand.arguments + [no_isolation_option]
+    arguments = BaseCommand.arguments + [lockfile_option, no_isolation_option]
 
     def add_arguments(self, parser: argparse.ArgumentParser) -> None:
         parser.add_argument(
