+++ pdm/cli/commands/remove.py
@@ -2,7 +2,7 @@ import argparse
 
 from pdm.cli import actions
 from pdm.cli.commands.base import BaseCommand
-from pdm.cli.options import dry_run_option, install_group
+from pdm.cli.options import dry_run_option, install_group, lockfile_option
 from pdm.project import Project
 
 
@@ -12,6 +12,7 @@ class Command(BaseCommand):
     def add_arguments(self, parser: argparse.ArgumentParser) -> None:
         install_group.add_to_parser(parser)
         dry_run_option.add_to_parser(parser)
+        lockfile_option.add_to_parser(parser)
         parser.add_argument(
             "-d",
             "--dev",
