+++ tests/cli/test_others.py
@@ -7,16 +7,6 @@ from pdm.utils import cd
 from tests import FIXTURES
 
 
-@pytest.mark.usefixtures("repository")
-def test_lock_dependencies(project):
-    project.add_dependencies({"requests": parse_requirement("requests")})
-    actions.do_lock(project)
-    assert project.lockfile_file.exists()
-    locked = project.locked_repository.all_candidates
-    for package in ("requests", "idna", "chardet", "certifi"):
-        assert package in locked
-
-
 @pytest.mark.usefixtures("project_no_init", "local_finder")
 def test_build_distributions(tmp_path, core):
     project = core.create_project()
@@ -49,12 +39,6 @@ def test_init_validate_python_requires(project_no_init):
 def test_help_option(invoke):
     result = invoke(["--help"])
     assert "PDM - Python Development Master" in result.output
-
-
-def test_lock_command(project, invoke, mocker):
-    m = mocker.patch.object(actions, "do_lock")
-    invoke(["lock"], obj=project)
-    m.assert_called_with(project, refresh=False)
 
 
 def test_info_command(project, invoke):
@@ -259,39 +243,6 @@ def test_completion_command(invoke):
     result = invoke(["completion", "bash"])
     assert result.exit_code == 0
     assert "(completion)" in result.output
-
-
-def test_lock_refresh(invoke, project, repository):
-    project.add_dependencies({"requests": parse_requirement("requests")})
-    result = invoke(["lock"], obj=project)
-    assert result.exit_code == 0
-    assert project.is_lockfile_hash_match()
-    assert not project.lockfile["metadata"]["files"].get("requests 2.19.1")
-    project.add_dependencies({"requests": parse_requirement("requests>=2.0")})
-    repository.get_hashes = (
-        lambda c: {"requests-2.19.1-py3-none-any.whl": "sha256:abcdef123456"}
-        if c.identify() == "requests"
-        else {}
-    )
-    assert not project.is_lockfile_hash_match()
-    result = invoke(["lock", "--refresh"], obj=project)
-    assert result.exit_code == 0
-    assert project.is_lockfile_hash_match()
-    assert project.lockfile["metadata"]["files"]["requests 2.19.1"][0] == {
-        "file": "requests-2.19.1-py3-none-any.whl",
-        "hash": "sha256:abcdef123456",
-    }
-
-
-def test_lock_refresh_keep_consistent(invoke, project, repository):
-    project.add_dependencies({"requests": parse_requirement("requests")})
-    result = invoke(["lock"], obj=project)
-    assert result.exit_code == 0
-    assert project.is_lockfile_hash_match()
-    previous = project.lockfile_file.read_text()
-    result = invoke(["lock", "--refresh"], obj=project)
-    assert result.exit_code == 0
-    assert project.lockfile_file.read_text() == previous
 
 
 @pytest.mark.network
